aws/awsContainer/code/requirements.txt
torch
transformers
pillow
datasets 
sentencepiece